Windows support (#34)
* Fix as std::isnan does not cast arg to double.
On windows cmath.h implementation std::isnan expects a double.
static_cast<double> added to resolve compilation error
C2668 'fpclassify': ambiguous call to overloaded function
* copts flag "-Wno-sign-compare" is not supported on windows.
Added windows config and selected copts accordingly.
* Revert designated initializers.
c++17 mode on cl.exe Version 19.26.28806 for x64 does not support designated initializers.
Reverted to c++17 compatible code.
* Added windows compatible unistd.h functions.
Added ifdef's to enable compilation on both windows and Unix systems.

+++ cc/algorithms/bounded-sum.h
@@ -92,7 +92,7 @@ class BoundedSum : public Algorithm<T> {
   };
 
   void AddEntry(const T& t) override {
-    if (std::isnan(t)) {
+    if (std::isnan(static_cast<double>(t))) {
       return;
     }
